pyproject.toml
[tool.poetry]
name = "exchange-api-test-task"
version = "1.1.1"
description = ""
authors = ["Nick Smirnov <na.smirnov@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
fastapi = "^0.115.6"
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
redis = "^5.2.0"
uvicorn = "^0.32.1"
httpx = "^0.28.0"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
ruff = "^0.8.1"
pylint = "^3.3.2"
flake8 = "^7.1.1"
mypy = "^1.13.0"
types-pyyaml = "^6.0.12.20240917"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
